Splitted the code into separate modules
