+++ index.html
@@ -2,15 +2,9 @@
 <html>
   <head>
     <meta charset="utf-8">
-
     <title>Web Engineering Coding Playground</title>
     <link href="https://fonts.googleapis.com/css?family=Open+Sans+Condensed:300|Sonsie+One" rel="stylesheet" type="text/css">
     <link rel="stylesheet" href="style.css">
-
-    <!-- the below three lines are a fix to get HTML5 semantic elements working in old versions of Internet Explorer-->
-    <!--[if lt IE 9]>
-      <script src="http://html5shiv.googlecode.com/svn/trunk/html5.js"></script>
-    <![endif]-->
   </head>
 
   <body>
@@ -116,7 +110,6 @@
           <div class="show-hide">Show comments</div>
 
           <div class="comment-wrapper">
-
             <font size="6">Add comment</font>
 
             <form class="comment-form">
@@ -136,29 +129,22 @@
             <font size="6">Comments</font>
 
             <ul class="comment-container">
-
               <li>
-
                 <p>Bob Fossil</p>
-
-                <p>Oh I am so glad you taught me all about the big brown angry guys in the woods. With their sniffing little noses and their bad attitudes, they can sure be a menace — I was thinking of putting them all in a truck and driving them outta here. I run a zoo, you know.</p>
-
+                <p>Oh I am so glad you taught me all about the big brown angry guys in the woods...</p>
               </li>
-
             </ul>
-
           </div>
         </section>
 
         <section class="more_bears">
           <font size="5">More Bears</font>
-
+          <!-- The bear data will be dynamically inserted here -->
         </section>
       </article>
 
       <div class="secondary">
         <font size="6">Related</font>
-
         <ul>
           <li><a href="#">The trouble with Bees</a></li>
           <li><a href="#">The trouble with Otters</a></li>
@@ -173,7 +159,7 @@
       <p>©Copyright 2050 by nobody. All rights reversed.</p>
     </footer>
 
-    <script src="main.js"></script>
-
+    <!-- Script tag for modules -->
+    <script src="main.js" type="module"></script>
   </body>
 </html>
